Move CircularList to common (tests in ui/ need)
