+++ src/Linkifier.test.ts
@@ -8,7 +8,7 @@ import { IMouseZoneManager, IMouseZone } from './ui/Types';
 import { ILinkMatcher, LineData, ITerminal } from './Types';
 import { Linkifier } from './Linkifier';
 import { MockBuffer, MockTerminal } from './utils/TestUtils.test';
-import { CircularList } from './utils/CircularList';
+import { CircularList } from './common/CircularList';
 
 class TestLinkifier extends Linkifier {
   constructor(terminal: ITerminal) {
