+++ src/common/CircularList.ts
